fix(test): move `fakeWindowForScript`
