+++ main_test.ts
@@ -2,19 +2,6 @@ import { expect } from "@std/expect";
 import { describe, it } from "@std/testing/bdd";
 
 import puppeteer from "puppeteer";
-import { HTMLScriptElement, Window } from "happy-dom";
-
-export const fakeWindowForScript = (scriptContent: string) => {
-  const window = new Window();
-  const document = window.document;
-  const script = document.createElement("script");
-  script.textContent = scriptContent;
-
-  const appendScript = (scriptToAppend?: HTMLScriptElement) =>
-    document.head.appendChild(scriptToAppend ?? script);
-
-  return { window, document, script, appendScript };
-};
 
 describe("PWA registration", () => {
   it("has registered service worker", async () => {
